+++ backend/packages/api/src/api/extension_management.py
