+++ backend/packages/cyris/pyproject.toml
@@ -8,7 +8,10 @@ dependencies = [
     "backend-commons",
     "litellm>=1.58.2",
     "rich>=13.9.2",
+    "async-generator>=1.10", # needed for ollama via litellm: https://docs.litellm.ai/docs/providers/ollama#example-usage---streaming--acompletion
     "utils",
+    "aiofiles>=24.1.0",
+    "types-aiofiles>=24.1.0.20241221",
 ]
 
 [build-system]
